\name{frm-package}
\alias{frm-package}
\docType{package}
\title{Regression analysis of fractional responses}
\description{This package estimates one- and two-part fractional regression models, tests the models' specification and calculates partial effects.
}
\details{
\tabular{ll}{
Package: \tab frm\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2014-01-29\cr
License: \tab GPL-2\cr
}
}
\author{Joaquim J.S. Ramalho <jsr@uevora.pt>}
\references{Ramalho, E.A., J.J.S. Ramalho and J.M.R. Murteira (2011),
"Alternative estimating and testing empirical strategies for fractional
regression models", \emph{Journal of Economic Surveys}, 25(1), 19-68.
}
