\name{frm-package}
\alias{frm-package}
\docType{package}
\title{Regression Analysis of Fractional Responses}
\description{Estimation and specification analysis of one- and two-part fractional regression models and calculation of partial effects.
}
\details{
\tabular{ll}{
Package: \tab frm\cr
Type: \tab Package\cr
Version: \tab 1.2.2\cr
Date: \tab 2015-02-20\cr
License: \tab GPL-2\cr
}
}
\author{Joaquim J.S. Ramalho <jsr@uevora.pt>}
\references{Ramalho, E.A., J.J.S. Ramalho and J.M.R. Murteira (2011),
"Alternative estimating and testing empirical strategies for fractional
regression models", \emph{Journal of Economic Surveys}, 25(1), 19-68.
}
